# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JustIdeas - Romanian Company Registration Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{id}/shareholders:
    get:
      operationId: listShareholders
      tags: [shareholders]
      summary: List shareholders of a company
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: List of shareholders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shareholder"
        "404":
          description: Company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: addShareholder
      tags: [shareholders]
      summary: Add a shareholder to a company
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShareholderBody"
      responses:
        "201":
          description: Shareholder added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shareholder"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Company not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /companies/{id}/shareholders/{shareholderId}:
    patch:
      operationId: updateShareholder
      tags: [shareholders]
      summary: Update a shareholder
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: shareholderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShareholderBody"
      responses:
        "200":
          description: Shareholder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shareholder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: removeShareholder
      tags: [shareholders]
      summary: Remove a shareholder from a company
      parameters:
        - $ref: "#/components/parameters/IdParam"
        - name: shareholderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Shareholder removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Directors ──────────────────────────────────────────────────────────────
...
  /registrations:
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List all registration requests
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, submitted, under_review, approved, rejected]
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of registration requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistrationRequest"

    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Create a registration request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegistrationBody"
      responses:
        "201":
          description: Registration request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{id}:
    get:
      operationId: getRegistration
      tags: [registrations]
      summary: Get a registration request by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Registration request found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateRegistration
      tags: [registrations]
      summary: Update a registration request
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRegistrationBody"
      responses:
        "200":
          description: Registration updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteRegistration
      tags: [registrations]
      summary: Delete a registration request
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Registration deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{id}/submit:
    post:
      operationId: submitRegistration
      tags: [registrations]
      summary: Submit a registration request for review
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Registration submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
        "400":
          description: Cannot submit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{id}/approve:
    post:
      operationId: approveRegistration
      tags: [registrations]
      summary: Approve a registration request
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRegistrationBody"
      responses:
        "200":
          description: Registration approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
        "400":
          description: Cannot approve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{id}/reject:
    post:
      operationId: rejectRegistration
      tags: [registrations]
      summary: Reject a registration request
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRegistrationBody"
      responses:
        "200":
          description: Registration rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationRequest"
        "400":
          description: Cannot reject
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Company ──────────────────────────────────────────────────────────────
...
    Shareholder:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        fullName:
          type: string
          description: Full name of the shareholder
        cnp:
          type: ["string", "null"]
          description: Personal identification number (CNP) for individuals
        idType:
          type: string
          enum: [ci, passport, company]
          description: ID document type
        idNumber:
          type: ["string", "null"]
          description: ID document number
        nationality:
          type: string
        address:
          type: ["string", "null"]
        numberOfShares:
          type: integer
          description: Number of shares held
        sharePercentage:
          type: number
          description: Percentage of shares owned
        contribution:
          type: number
          description: Capital contribution in RON
        isLegalEntity:
          type: boolean
          description: True if the shareholder is a company, false if individual
        companyRegistrationNumber:
          type: ["string", "null"]
          description: Registration number if shareholder is a legal entity
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - fullName
        - idType
        - nationality
        - numberOfShares
        - sharePercentage
        - contribution
        - isLegalEntity
        - createdAt
        - updatedAt
    CreateShareholderBody:
      type: object
      properties:
        fullName:
          type: string
        cnp:
          type: ["string", "null"]
        idType:
          type: string
          enum: [ci, passport, company]
        idNumber:
          type: ["string", "null"]
        nationality:
          type: string
        address:
          type: ["string", "null"]
        numberOfShares:
          type: integer
        sharePercentage:
          type: number
        contribution:
          type: number
        isLegalEntity:
          type: boolean
        companyRegistrationNumber:
          type: ["string", "null"]
      required:
        - fullName
        - idType
        - nationality
        - numberOfShares
        - sharePercentage
        - contribution
        - isLegalEntity
    UpdateShareholderBody:
      type: object
      properties:
        fullName:
          type: string
        cnp:
          type: ["string", "null"]
        idType:
          type: string
          enum: [ci, passport, company]
        idNumber:
          type: ["string", "null"]
        nationality:
          type: string
        address:
          type: ["string", "null"]
        numberOfShares:
          type: integer
        sharePercentage:
          type: number
        contribution:
          type: number
        isLegalEntity:
          type: boolean
        companyRegistrationNumber:
          type: ["string", "null"]

    # ── Director ──────────────────────────────────────────────────────────────
...
    RegistrationRequest:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, submitted, under_review, approved, rejected]
        submittedAt:
          type: ["string", "null"]
          format: date-time
        reviewedAt:
          type: ["string", "null"]
          format: date-time
        reviewerNotes:
          type: ["string", "null"]
        estimatedCompletionDate:
          type: ["string", "null"]
          format: date
        referenceNumber:
          type: ["string", "null"]
          description: Official reference number assigned after submission
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - status
        - createdAt
        - updatedAt
    CreateRegistrationBody:
      type: object
      properties:
        companyId:
          type: integer
        estimatedCompletionDate:
          type: ["string", "null"]
          format: date
      required:
        - companyId
    UpdateRegistrationBody:
      type: object
      properties:
        estimatedCompletionDate:
          type: ["string", "null"]
          format: date
        reviewerNotes:
          type: ["string", "null"]
        referenceNumber:
          type: ["string", "null"]
    ReviewRegistrationBody:
      type: object
      properties:
        reviewerNotes:
          type: ["string", "null"]
        referenceNumber:
          type: ["string", "null"]
        estimatedCompletionDate:
          type: ["string", "null"]
          format: date