# lib/api-zod/src/generated/types/createRegistrationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JustIdeas - Romanian Company Registration Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRegistrationBody {
  companyId: number;
  /** @nullable */
  estimatedCompletionDate?: Date | null;
}