# lib/api-zod/src/generated/types/registrationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JustIdeas - Romanian Company Registration Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationRequestStatus } from "./registrationRequestStatus";

export interface RegistrationRequest {
  id: number;
  companyId: number;
  /** @nullable */
  companyName?: string | null;
  status: RegistrationRequestStatus;
  /** @nullable */
  submittedAt?: Date | null;
  /** @nullable */
  reviewedAt?: Date | null;
  /** @nullable */
  reviewerNotes?: string | null;
  /** @nullable */
  estimatedCompletionDate?: Date | null;
  /**
   * Official reference number assigned after submission
   * @nullable
   */
  referenceNumber?: string | null;
  createdAt: Date;
  updatedAt: Date;
}