# lib/api-zod/src/generated/types/reviewRegistrationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JustIdeas - Romanian Company Registration Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ReviewRegistrationBody {
  /** @nullable */
  reviewerNotes?: string | null;
  /** @nullable */
  referenceNumber?: string | null;
  /** @nullable */
  estimatedCompletionDate?: Date | null;
}